* Standardise handling of MIDI Pitch Bend messages from all backends. * Fix failure to apply MIDI mappings on load. * Fix failure to open ALSA MIDI device on load. * MIDI Map: <translations> Static mapping of MIDI values per controller (use for eg configuring an expanded "zero point") * Support for M-Audio Xponent MIDI controller, including wheel touch sensors to swap between jog and scratch mode. In scratch mode, temporarily disable pitch-independent timestretch for a more vinyl-like sound. * Spread64 midioption (aka Relative CC) with sensitivity adjustment, for M-Audio Xponent jogwheels and maybe other controllers.

+++ mixxx/src/visual/visualbuffer.cpp
@@ -47,7 +47,7 @@ VisualBuffer::VisualBuffer(EngineBuffer *pEngineBuffer, const char *group)
     m_fDisplayFactor = 1.;
     
     // Length of this buffer. 
-    m_iLen = MAXDISPLAYRATE*READBUFFERSIZE/(m_pSampleRate->get());
+    m_iLen = MAXDISPLAYRATE*READBUFFERSIZE/(int)m_pSampleRate->get();
     if (!even(m_iLen))
         m_iLen--;
     m_iSourceLen = m_iLen;
